fix MSVC warning

+++ extras/minimize_testcase.cpp
@@ -84,7 +84,7 @@ bool safeToRemove(const Polygons &polys, size_t i, size_t j, float precision) {
   return true;
 }
 
-std::pair<int, int> findIndex(const Polygons &polys, size_t i) {
+std::pair<size_t, size_t> findIndex(const Polygons &polys, size_t i) {
   size_t outer = 0;
   while (i >= polys[outer].size()) i -= polys[outer++].size();
   return std::make_pair(outer, i);
@@ -113,9 +113,9 @@ void DumpTriangulation(const Polygons &polys, float precision) {
   std::vector<glm::ivec3> result = Triangulate(polys);
   PolygonParams() = oldParams;
   for (const glm::ivec3 &tri : result) {
-    std::pair<int, int> xid = findIndex(polys, tri.x);
-    std::pair<int, int> yid = findIndex(polys, tri.y);
-    std::pair<int, int> zid = findIndex(polys, tri.z);
+    std::pair<size_t, size_t> xid = findIndex(polys, tri.x);
+    std::pair<size_t, size_t> yid = findIndex(polys, tri.y);
+    std::pair<size_t, size_t> zid = findIndex(polys, tri.z);
     glm::vec2 x = polys[xid.first][xid.second];
     glm::vec2 y = polys[yid.first][yid.second];
     glm::vec2 z = polys[zid.first][zid.second];
